Improve /reserve endpoint
The /reserve endpoint handles POST request and marks a booked table as reserved. When a table is reserved a unique token is generated and given to the user as prove of the reservation made.

The BookedTable class is added.

When making a reserve the table is marked as booked and is retired form the available tables. At the moment this feature is down.

The endpoint works as follows:

Request body:
{
    "tableNumber": 1
}

if the reserve is successful, the response body is as follows:
{
    "token": "ae4c53d7-a39a-42a0-ab52-f3bd1b847f38",
    "table": {
        "tableNumber": 1,
        "numberOfSeats": 6
    }
}

if the table is taken:
{
    "error": "The table has been already reserved."
}

and if the table number is out of bounds or is a negative number:
{
    "error": "The table number is out of bounds!."
}