Implement /return endpoint
The /return endpoint handles POST request and undo a previous table reservation. The request should have the token that identifies the table reservation, with the return token the table it relates to is identified and is marked as available.

The endpoint works as follows:

Request body:
{
    "token": "ae4c53d7-a39a-42a0-ab52-f3bd1b847f38"
}

if the token is correct the response body is as follows:
{
    "returnedTable": {
        "tableNumber": 3,
        "numberOfSeats": 6
    }
}

response body with an expired token
{
    "error": ""Wrong token!""
}