Fix /reserve /return endpoint
Fix the functionality of both /reserve and /return endpoints.

When a successful /"reserve" call is made, the reserved table is taking down from the available table list, and isn't shown in the /"tables" call.
When a successful /"return" call is made, the returned table is identified and added to the available table list.